MCR-3498 Moved XMLCleaner classes to separate package
