+++ mycore-xeditor/src/main/java/org/mycore/frontend/xeditor/cleanup/MCRXMLCleaner.java
@@ -16,7 +16,7 @@
  * along with MyCoRe.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.mycore.frontend.xeditor;
+package org.mycore.frontend.xeditor.cleanup;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -27,10 +27,6 @@ import java.util.Map;
 import org.jdom2.Attribute;
 import org.jdom2.Document;
 import org.jdom2.Element;
-import org.jdom2.filter.Filters;
-import org.jdom2.xpath.XPathExpression;
-import org.jdom2.xpath.XPathFactory;
-import org.mycore.common.MCRConstants;
 
 public class MCRXMLCleaner {
 
@@ -106,49 +102,5 @@ public class MCRXMLCleaner {
     private boolean isRelevant(Object node) {
         MCRCleaningRule rule = nodes2rules.get(node);
         return (rule == null || rule.isRelevant(node));
-    }
-}
-
-class MCRCleaningRule {
-
-    private String xPathExprNodesToInspect;
-
-    private XPathExpression<Object> xPathNodesToInspect;
-
-    private XPathExpression<Object> xPathRelevancyTest;
-
-    MCRCleaningRule(String xPathExprNodesToInspect, String xPathExprRelevancyTest) {
-        this.xPathExprNodesToInspect = xPathExprNodesToInspect;
-        this.xPathNodesToInspect = XPathFactory.instance().compile(xPathExprNodesToInspect, Filters.fpassthrough(),
-            null,
-            MCRConstants.getStandardNamespaces());
-        this.xPathRelevancyTest = XPathFactory.instance().compile(xPathExprRelevancyTest, Filters.fpassthrough(), null,
-            MCRConstants.getStandardNamespaces());
-    }
-
-    public List<Object> getNodesToInspect(Document xml) {
-        return xPathNodesToInspect.evaluate(xml);
-    }
-
-    public boolean isRelevant(Object node) {
-        Object found = xPathRelevancyTest.evaluateFirst(node);
-        if (found == null) {
-            return false;
-        } else if (found instanceof Boolean b) {
-            return b;
-        } else {
-            return true; // something matching found
-        }
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        return obj instanceof MCRCleaningRule cleaningRule
-            && xPathExprNodesToInspect.equals(cleaningRule.xPathExprNodesToInspect);
-    }
-
-    @Override
-    public int hashCode() {
-        return xPathExprNodesToInspect.hashCode();
     }
 }
